+++ plugins/org.jkiss.dbeaver.ext.cubrid/plugin.xml
@@ -62,8 +62,6 @@
                                         </items>
                                     </items>
                                 </folder>
-                                <items label="%tree.sub_table.node.name" optional="true" path="subtable" property="subTables" icon="#table" recursive=".." visibleIf="!object.view">
-                                </items>
                             </items>
                             <folder type="org.jkiss.dbeaver.ext.cubrid.model.CubridTable" label="%tree.system_tables.node.name" icon="#folder_table" description="%tree.system_table.node.tip" visibleIf="object.supportsSystemTable() &amp;&amp; object.showSystemTableFolder()">
                                 <items label="%tree.table.node.name" path="table" property="physicalSystemTables" icon="#table">
@@ -95,8 +93,6 @@
                                             </items>
                                         </items>
                                     </folder>
-                                    <items label="%tree.sub_table.node.name" optional="true" path="subtable" property="subTables" icon="#table" recursive=".." visibleIf="!object.view">
-                                    </items>
                                 </items>
                             </folder>
                         </folder>
@@ -115,6 +111,12 @@
                                     </folder>
                                 </items>
                             </folder>
+                        </folder>
+                        <folder type="org.jkiss.dbeaver.ext.generic.model.GenericTableIndex" label="%tree.indexes.node.name" icon="#indexes" description="%tree.indexes.node.tip">
+                            <items label="%tree.index.node.name" path="index" property="indexes" icon="#index">
+                                <items label="%tree.index_columns.node.name" path="column" property="attributeReferences" icon="#column" navigable="false" inline="true">
+                                </items>
+                            </items>
                         </folder>
                     </items>
                 </folder>
