+++ plugins/org.jkiss.dbeaver.ext.cubrid/src/org/jkiss/dbeaver/ext/cubrid/model/CubridDataSource.java
@@ -35,9 +35,7 @@ import java.util.Map;
 
 public class CubridDataSource extends GenericDataSource
 {
-
     private final CubridMetaModel metaModel;
-    private final CubridObjectContainer structureContainer;
     private boolean supportMultiSchema;
 
     public CubridDataSource(DBRProgressMonitor monitor, DBPDataSourceContainer container, CubridMetaModel metaModel)
@@ -45,7 +43,6 @@ public class CubridDataSource extends GenericDataSource
     {
         super(monitor, container, metaModel, new CubridSQLDialect());
         this.metaModel = new CubridMetaModel();
-        this.structureContainer = new CubridObjectContainer(this);
     }
 
     @DPIContainer
@@ -92,11 +89,6 @@ public class CubridDataSource extends GenericDataSource
             types.put(dataType.getName(), dataType);
         }
         return types.values();
-    }
-
-    public CubridObjectContainer getObjectContainer()
-    {
-        return structureContainer;
     }
 
     @Override
